Remove duplicate code in deserializer
